requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pdfplumber==0.11.4
python-docx==1.1.2
openai==1.55.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
